Fixed SomeThings
